+++ src/Navbar/Navbar.tsx
@@ -1,0 +1,111 @@
+import * as React from 'react';
+import AppBar from '@mui/material/AppBar';
+import Box from '@mui/material/Box';
+import Toolbar from '@mui/material/Toolbar';
+import Typography from '@mui/material/Typography';
+import IconButton from '@mui/material/IconButton';
+import MenuIcon from '@mui/icons-material/Menu';
+import Container from '@mui/material/Container';
+import MenuItem from '@mui/material/MenuItem';
+import Menu from '@mui/material/Menu';
+import { Link } from "react-router-dom";
+
+import { WalletSelector } from "@aptos-labs/wallet-adapter-ant-design"; 
+import "@aptos-labs/wallet-adapter-ant-design/dist/index.css";
+
+import gatosLogo from '../assets/images/GATOS.png';
+
+export default function Navbar() {
+    const [anchorEl, setAnchorEl] = React.useState<null | HTMLElement>(null);
+  const [mobileMoreAnchorEl, setMobileMoreAnchorEl] =
+    React.useState<null | HTMLElement>(null);
+
+  const open = Boolean(anchorEl);
+  const handleClick = (event: React.MouseEvent<HTMLElement>) => {
+    setAnchorEl(event.currentTarget);
+  };
+  const handleClose = () => {
+    setAnchorEl(null);
+  };
+
+  const isMenuOpen = Boolean(anchorEl);
+  const isMobileMenuOpen = Boolean(mobileMoreAnchorEl);
+
+    return (
+        <div>
+            <Container maxWidth="xl">
+            <AppBar position="static"
+                sx={{backgroundColor:'black'}}>
+                <Toolbar>
+                <IconButton
+                    size="large"
+                    edge="start"
+                    color="inherit"
+                    aria-label="menu"
+                    sx={{ mr: 2 }}
+                    onClick={handleClick}
+                >
+                    <MenuIcon sx={{color:'#0ACFFE'}}/>
+                    <Menu
+                anchorEl={anchorEl}
+                id="account-menu"
+                open={open}
+                onClose={handleClose}
+                onClick={handleClose}
+                PaperProps={{
+                elevation: 0,
+                sx: {
+                    overflow: 'visible',
+                    filter: 'drop-shadow(0px 2px 8px rgba(0,0,0,0.32))',
+                    mt: 1.5,
+                    '& .MuiAvatar-root': {
+                    width: 32,
+                    height: 32,
+                    ml: -0.5,
+                    mr: 1,
+                    },
+                    '&:before': {
+                    content: '""',
+                    display: 'block',
+                    position: 'absolute',
+                    top: 0,
+                    right: 14,
+                    width: 10,
+                    height: 10,
+                    bgcolor: 'background.paper',
+                    transform: 'translateY(-50%) rotate(45deg)',
+                    zIndex: 0,
+                    },
+                },
+                }}
+                transformOrigin={{ horizontal: 'right', vertical: 'top' }}
+                anchorOrigin={{ horizontal: 'right', vertical: 'bottom' }}
+            >
+                <MenuItem onClick={handleClose}>
+                <Link to="/">Home</Link>
+                </MenuItem>
+                <MenuItem onClick={handleClose}>
+                <Link to="/gamefeed">GameFeed</Link>
+                </MenuItem>
+                <MenuItem onClick={handleClose}>
+                <Link to="/mygames">MyGames</Link>
+                </MenuItem>
+                <MenuItem onClick={handleClose}>
+                <Link to="/community">Community</Link>
+                </MenuItem>
+                <MenuItem onClick={handleClose}>
+                <Link to="/test">Test</Link>
+                </MenuItem>
+            </Menu>
+                </IconButton>
+                <div>
+                    <img src={gatosLogo} alt="logo" width="80vw"/>
+                </div>
+                <Box sx={{width:'500vw'}}></Box>
+                <WalletSelector/>
+                </Toolbar>
+            </AppBar>
+            </Container>
+        </div>
+    );
+}
